requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.3
flask-cors==6.0.2
gunicorn==25.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
numpy==2.4.2
packaging==26.0
pandas==3.0.1
PuLP==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.3
Werkzeug==3.1.6
xgboost==3.2.0
google-genai
openai